+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added DSC_VirtualHardDisk resource for creating virtual disks and tests.
+
 ## [5.1.0] - 2023-02-22
 
 ### Changed